requirements.txt
Flask
Flask-Babel
numpy
Pillow
tensorflow
